IT MATE PK/step15 part 01 (runner methods) streaming/pyproject.toml
[project]
name = "step00"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "openai-agents>=0.2.11",
    "python-decouple>=3.8",
    "rich>=14.1.0",
]
